+++ React/src/components/sections/HospitalTable.js
@@ -1,82 +1,430 @@
 import React from "react";
-import { Table, Thead, Tbody, Tr, Th, Td } from "react-super-responsive-table";
-import "react-super-responsive-table/dist/SuperResponsiveTableStyle.css";
-import axios from "axios";
-import { Pagination } from "react-bootstrap";
-import "bootstrap/dist/css/bootstrap.min.css";
 
-export default class table extends React.Component {
-  constructor() {
-    super();
-    this.state = {
-      data: [],
-      loading: false,
-      sort: "none",
-      totalItems: 60,
-      itemsPerPage: 10,
-      currentPageNumber: 1,
-    };
-  }
+import PropTypes from "prop-types";
+import clsx from "clsx";
+import { lighten, makeStyles, darken } from "@material-ui/core/styles";
+import Table from "@material-ui/core/Table";
+import TableBody from "@material-ui/core/TableBody";
+import TableCell from "@material-ui/core/TableCell";
+import TableContainer from "@material-ui/core/TableContainer";
+import TableHead from "@material-ui/core/TableHead";
+import TablePagination from "@material-ui/core/TablePagination";
+import TableRow from "@material-ui/core/TableRow";
+import TableSortLabel from "@material-ui/core/TableSortLabel";
+import Toolbar from "@material-ui/core/Toolbar";
+import Typography from "@material-ui/core/Typography";
+import Paper from "@material-ui/core/Paper";
+//import "react-super-responsive-table/dist/SuperResponsiveTableStyle.css";
+//import "bootstrap/dist/css/bootstrap.min.css";
+//import Checkbox from "@material-ui/core/Checkbox";
+import IconButton from "@material-ui/core/IconButton";
+import Tooltip from "@material-ui/core/Tooltip";
+import FormControlLabel from "@material-ui/core/FormControlLabel";
+import Switch from "@material-ui/core/Switch";
+import DeleteIcon from "@material-ui/icons/Delete";
+import FilterListIcon from "@material-ui/icons/FilterList";
+import { textAlign } from "@material-ui/system";
 
-  componentDidMount() {
-    this.setState({ loading: true });
-    axios
-      .get(
-        `https://iphy6zyd24.execute-api.ap-south-1.amazonaws.com/test1/hospital`
-      )
-      .then((res) => {
-        const data = res.data;
-        this.setState({
-          data,
-          currentPageNumber: res.currentPageNumber,
-          totalItems: res.totalItems,
-          itemsPerPage: res.itemsPerPage,
-        });
+function descendingComparator(a, b, orderBy) {
+  if (b[orderBy] < a[orderBy]) {
+    return -1;
+  }
+  if (b[orderBy] > a[orderBy]) {
+    return 1;
+  }
+  return 0;
+}
 
-        // console.log("data");
-        // console.log(data);
-      });
-    this.setState({ loading: false });
-  }
+function getComparator(order, orderBy) {
+  return order === "desc"
+    ? (a, b) => descendingComparator(a, b, orderBy)
+    : (a, b) => -descendingComparator(a, b, orderBy);
+}
 
-  handleSelect(number) {
-    console.log("handle select", number);
-  }
-  render() {
-    const rows = this.state.data;
-    return (
-      <div style={{ position: "relative", margin: "20px" }}>
-        <Table>
-          <Thead>
-            <Tr>
-              <Th>Serial No.</Th>
-              <Th>Hospital Name</Th>
-              <Th>Covid Beds</Th>
-              <Th>Covid Ventilators</Th>
-              <Th>Distance (Kms)</Th>
-              <Th>Address</Th>
-              <Th>Contact</Th>
-            </Tr>
-          </Thead>
-          <Tbody>
-            {rows.map((row) => (
-              <Tr>
-                <Td>{row.serialNo}</Td>
-                <Td>{row.hospital}</Td>
-                {row.covidBeds > 0 ? <Td>{row.covidBeds}</Td> : <Td>0</Td>}
-                {row.covidVentillator > 0 ? (
-                  <Td>{row.covidVentillator}</Td>
-                ) : (
-                  <Td>0</Td>
-                )}
-                <Td>{row.distance}</Td>
-                <Td>{row.address}</Td>
-                <Td>{row.mobileNo}</Td>
-              </Tr>
-            ))}
-          </Tbody>
-        </Table>
-      </div>
-    );
-  }
+function stableSort(array, comparator) {
+  const stabilizedThis = array.map((el, index) => [el, index]);
+  stabilizedThis.sort((a, b) => {
+    const order = comparator(a[0], b[0]);
+    if (order !== 0) return order;
+    return a[1] - b[1];
+  });
+  return stabilizedThis.map((el) => el[0]);
+}
+
+const headCells = [
+  {
+    id: "hospital",
+    numeric: false,
+    disablePadding: true,
+    label: "HOSPITAL NAME",
+  },
+  {
+    id: "covidBeds",
+    numeric: true,
+    disablePadding: true,
+    label: "COVID BEDS",
+  },
+  {
+    id: "covidVentillator",
+    numeric: true,
+    disablePadding: false,
+    label: "COVID VENTILATORS",
+  },
+  {
+    id: "distance",
+    numeric: false,
+    disablePadding: true,
+    label: "DISTANCE (Kms)",
+  },
+  {
+    id: "address",
+    numeric: false,
+    disablePadding: false,
+    label: "ADDRESS",
+  },
+  {
+    id: "contact",
+    numeric: true,
+    disablePadding: false,
+    label: "CONTACT",
+  },
+];
+
+function EnhancedTableHead(props) {
+  const {
+    classes,
+    onSelectAllClick,
+    order,
+    orderBy,
+    numSelected,
+    rowCount,
+    onRequestSort,
+  } = props;
+  const createSortHandler = (property) => (event) => {
+    onRequestSort(event, property);
+  };
+
+  return (
+    <TableHead>
+      <TableRow>
+        {/* <TableCell padding="checkbox">
+          <Checkbox
+            indeterminate={numSelected > 0 && numSelected < rowCount}
+            checked={rowCount > 0 && numSelected === rowCount}
+            onChange={onSelectAllClick}
+            inputProps={{ "aria-label": "select all" }}
+          />
+        </TableCell> */}
+        {headCells.map((headCell) => (
+          <TableCell
+            key={headCell.id}
+            align={headCell.numeric ? "left" : "left"}
+            padding={headCell.disablePadding ? "none" : "default"}
+            sortDirection={orderBy === headCell.id ? order : false}
+          >
+            <TableSortLabel
+              active={orderBy === headCell.id}
+              direction={orderBy === headCell.id ? order : "desc"}
+              onClick={createSortHandler(headCell.id)}
+            >
+              {headCell.label}
+              {orderBy === headCell.id ? (
+                <span className={classes.visuallyHidden}>
+                  {order === "desc" ? "sorted descending" : "sorted ascending"}
+                </span>
+              ) : null}
+            </TableSortLabel>
+          </TableCell>
+        ))}
+      </TableRow>
+    </TableHead>
+  );
+}
+
+EnhancedTableHead.propTypes = {
+  classes: PropTypes.object.isRequired,
+  numSelected: PropTypes.number.isRequired,
+  onRequestSort: PropTypes.func.isRequired,
+  onSelectAllClick: PropTypes.func.isRequired,
+  order: PropTypes.oneOf(["asc", "desc"]).isRequired,
+  orderBy: PropTypes.string.isRequired,
+  rowCount: PropTypes.number.isRequired,
+};
+
+const useToolbarStyles = makeStyles((theme) => ({
+  root: {
+    paddingLeft: theme.spacing(2),
+    paddingRight: theme.spacing(1),
+  },
+  highlight:
+    theme.palette.type === "dark"
+      ? {
+          color: theme.palette.secondary.dark,
+          backgroundColor: darken(theme.palette.secondary.dark, 0.85),
+        }
+      : {
+          color: theme.palette.text.primary,
+          backgroundColor: theme.palette.secondary.dark,
+        },
+  title: {
+    flex: "1 1 100%",
+  },
+}));
+
+const EnhancedTableToolbar = (props) => {
+  const classes = useToolbarStyles();
+  const { numSelected } = props;
+
+  return (
+    <Toolbar
+      className={clsx(classes.root, {
+        [classes.highlight]: numSelected > 0,
+      })}
+    >
+      {numSelected > 0 ? (
+        <Typography
+          className={classes.title}
+          color="inherit"
+          variant="subtitle1"
+          component="div"
+        >
+          {numSelected} selected
+        </Typography>
+      ) : (
+        <Typography
+          className={classes.title}
+          variant="h4"
+          id="tableTitle"
+          component="div"
+          align="center"
+        >
+          {/* <br></br>
+          TESTING CENTERS */}
+        </Typography>
+      )}
+      <br></br>
+
+      {/* {numSelected > 0 ? (
+        <Tooltip title="Delete">
+          <IconButton aria-label="delete">
+            <DeleteIcon />
+          </IconButton>
+        </Tooltip>
+      ) : (
+        <Tooltip title="Filter list">
+          <IconButton aria-label="filter list">
+            <FilterListIcon />
+          </IconButton>
+        </Tooltip>
+      )} */}
+    </Toolbar>
+  );
+};
+
+EnhancedTableToolbar.propTypes = {
+  numSelected: PropTypes.number.isRequired,
+};
+
+const useStyles = makeStyles((theme) => ({
+  root: {
+    width: "100%",
+    paddingTop: theme.spacing(2),
+  },
+  paper: {
+    width: "100%",
+    marginBottom: theme.spacing(10),
+    paddingRight: theme.spacing(6),
+    paddingLeft: theme.spacing(6),
+  },
+  table: {
+    minWidth: 750,
+    //marginLeft: 30,
+    //marginRight: 30
+  },
+  visuallyHidden: {
+    border: 0,
+    clip: "rect(0 0 0 0)",
+    height: 1,
+    margin: -1,
+    overflow: "hidden",
+    padding: 10,
+    position: "absolute",
+    top: 20,
+    width: 1,
+  },
+}));
+
+export default function EnhancedTable(props) {
+  const rows = props.data;
+  console.log(rows);
+  const classes = useStyles();
+  const [order, setOrder] = React.useState("asc");
+  const [orderBy, setOrderBy] = React.useState("distance");
+  // const [orderBy1, setOrderBy1] = React.useState("covidVentillator");
+  // const [orderBy2, setOrderBy2] = React.useState("distance");
+  const [selected, setSelected] = React.useState([]);
+  const [page, setPage] = React.useState(0);
+  const [dense, setDense] = React.useState(false);
+  const [rowsPerPage, setRowsPerPage] = React.useState(10);
+
+  const handleRequestSort = (event, property) => {
+    const isAsc = orderBy === property && order === "asc";
+    setOrder(isAsc ? "desc" : "asc");
+    setOrderBy(property);
+    // setOrderBy1(property);
+    // setOrderBy2(property);
+  };
+
+  // const handleSelectAllClick = event => {
+  //   if (event.target.checked) {
+  //     const newSelecteds = rows.map(n => n.name);
+  //     setSelected(newSelecteds);
+  //     return;
+  //   }
+  //   setSelected([]);
+  // };
+
+  // const handleClick = (event, name) => {
+  //   const selectedIndex = selected.indexOf(name);
+  //   let newSelected = [];
+
+  //   if (selectedIndex === -1) {
+  //     newSelected = newSelected.concat(selected, name);
+  //   } else if (selectedIndex === 0) {
+  //     newSelected = newSelected.concat(selected.slice(1));
+  //   } else if (selectedIndex === selected.length - 1) {
+  //     newSelected = newSelected.concat(selected.slice(0, -1));
+  //   } else if (selectedIndex > 0) {
+  //     newSelected = newSelected.concat(
+  //       selected.slice(0, selectedIndex),
+  //       selected.slice(selectedIndex + 1)
+  //     );
+  //   }
+
+  //   setSelected(newSelected);
+  // };
+
+  const handleChangePage = (event, newPage) => {
+    setPage(newPage);
+  };
+
+  const handleChangeRowsPerPage = (event) => {
+    setRowsPerPage(parseInt(event.target.value, 10));
+    setPage(0);
+  };
+
+  const handleChangeDense = (event) => {
+    setDense(event.target.checked);
+  };
+
+  const isSelected = (name) => selected.indexOf(name) !== -1;
+
+  const emptyRows =
+    rowsPerPage - Math.min(rowsPerPage, rows.length - page * rowsPerPage);
+
+  return (
+    <div className={classes.root}>
+      <Paper className={classes.paper}>
+        <EnhancedTableToolbar numSelected={selected.length} />
+        <TableContainer>
+          <Table
+            className={classes.table}
+            aria-labelledby="tableTitle"
+            size={dense ? "small" : "medium"}
+            aria-label="enhanced table"
+          >
+            <EnhancedTableHead
+              classes={classes}
+              numSelected={selected.length}
+              order={order}
+              orderBy={orderBy}
+              // onSelectAllClick={handleSelectAllClick}
+              onRequestSort={handleRequestSort}
+              rowCount={rows.length}
+            />
+            <TableBody>
+              {stableSort(rows, getComparator(order, orderBy))
+                .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
+                .map((row, index) => {
+                  const isItemSelected = isSelected(row.name);
+                  const labelId = `enhanced-table-checkbox-${index}`;
+
+                  return (
+                    <TableRow
+                      hover
+                      // onClick={event => handleClick(event, row.name)}
+                      role="checkbox"
+                      aria-checked={isItemSelected}
+                      tabIndex={-1}
+                      key={row.name}
+                      selected={isItemSelected}
+                    >
+                      <TableCell
+                        component="th"
+                        id={labelId}
+                        scope="row"
+                        padding="none"
+                        align="left"
+                      >
+                        {row.hospital}
+                      </TableCell>
+
+                      <TableCell
+                        component="th"
+                        id={labelId}
+                        scope="row"
+                        padding="none"
+                        align="left"
+                      >
+                        {row.covidBeds}
+                      </TableCell>
+
+                      <TableCell
+                        component="th"
+                        id={labelId}
+                        scope="row"
+                        padding="none"
+                        align="center"
+                      >
+                        {row.covidVentillator}
+                      </TableCell>
+                      <TableCell
+                        component="th"
+                        id={labelId}
+                        scope="row"
+                        padding="none"
+                        align="left"
+                      >
+                        {row.distance}
+                      </TableCell>
+
+                      <TableCell align="left">{row.address}</TableCell>
+                      <TableCell align="left">{row.mobileNo}</TableCell>
+                    </TableRow>
+                  );
+                })}
+              {emptyRows > 0 && (
+                <TableRow style={{ height: (dense ? 33 : 53) * emptyRows }}>
+                  <TableCell colSpan={6} />
+                </TableRow>
+              )}
+            </TableBody>
+          </Table>
+        </TableContainer>
+        <TablePagination
+          rowsPerPageOptions={[10, 20, 35]}
+          component="div"
+          count={rows.length}
+          rowsPerPage={rowsPerPage}
+          page={page}
+          onChangePage={handleChangePage}
+          onChangeRowsPerPage={handleChangeRowsPerPage}
+        />
+      </Paper>
+      {/* <FormControlLabel
+        control={<Switch checked={dense} onChange={handleChangeDense} />}
+        label="Dense padding"
+      /> */}
+    </div>
+  );
 }
